requirements.txt
# 核心依赖
pytorch-lightning==2.1.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2

# 数据处理
numpy==1.24.3
pillow==10.0.0

# 配置解析
omegaconf==2.3.0
PyYAML==6.0.1

# 日志与可视化
tensorboard==2.14.0

# 模型库（分割任务示例）
segmentation-models-pytorch==0.3.3

# 工具
tqdm==4.65.0